MINOR: Move test specific ruby classloading hack to test code
