+++ logstash-core/src/main/java/org/logstash/config/ir/ConfigCompiler.java
@@ -1,11 +1,7 @@
 package org.logstash.config.ir;
 
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import org.jruby.RubyHash;
 import org.jruby.javasupport.JavaUtil;
 import org.jruby.runtime.builtin.IRubyObject;
-import org.jruby.runtime.load.LoadService;
 import org.logstash.RubyUtil;
 import org.logstash.common.IncompleteSourceWithMetadataException;
 import org.logstash.common.SourceWithMetadata;
@@ -28,7 +24,6 @@ public final class ConfigCompiler {
      */
     public static PipelineIR configToPipelineIR(final String config, final boolean supportEscapes)
         throws IncompleteSourceWithMetadataException {
-        ensureLoadpath();
         final IRubyObject compiler = RubyUtil.RUBY.executeScript(
             "require 'logstash/compiler'\nLogStash::Compiler",
             ""
@@ -46,24 +41,5 @@ public final class ConfigCompiler {
                 }
             );
         return (PipelineIR) code.toJava(PipelineIR.class);
-    }
-
-    /**
-     * Loads the logstash-core/lib path if the load service can't find {@code logstash/compiler}.
-     */
-    private static void ensureLoadpath() {
-        final LoadService loader = RubyUtil.RUBY.getLoadService();
-        if (loader.findFileForLoad("logstash/compiler").library == null) {
-            final RubyHash environment = RubyUtil.RUBY.getENV();
-            final Path root = Paths.get(
-                System.getProperty("logstash.core.root.dir", "")
-            ).toAbsolutePath();
-            final String gems = root.getParent().resolve("vendor").resolve("bundle")
-                .resolve("jruby").resolve("2.3.0").toFile().getAbsolutePath();
-            environment.put("GEM_HOME", gems);
-            environment.put("GEM_PATH", gems);
-            loader.addPaths(root.resolve("lib").toFile().getAbsolutePath()
-            );
-        }
     }
 }
